Last commit before switching to Hive because Isar web support is currently broken.

+++ lib/src/cached_data.dart
@@ -19,7 +19,7 @@ class CachedData {
   Id get id => stringToHash(storageRefFullPath);
 
   /// The bytes stored
-  final List<int> bytes;
+  final List<byte> bytes;
 
   /// Cache created
   final DateTime cacheCreated;
